+++ ChangeLog.md
@@ -4,6 +4,7 @@
 ### 2018-08-09  Thomas Deuling  <typo3@coding.ms>
 
 *   [TASK] Adding allowed content types for carousel element.
+*   [TASK] Changing integration of colPos in TCA of tt_content.
 
 ### 2018-08-07  Thomas Deuling  <typo3@coding.ms>